+++ src/apis/member/__init__.py
@@ -18,6 +18,7 @@ routes = [
     APIRoute("/member/roles/rank/{rank_type_id}", userop.patch_roles_rank, methods=["PATCH"], response_class=JSONResponse),
     APIRoute("/member/bonus/history", userop.get_bonus_history, methods=["GET"], response_class=JSONResponse),
     APIRoute("/member/bonus/claim", userop.post_bonus_claim, methods=["POST"], response_class=JSONResponse),
+    APIRoute("/member/bonus/notification/settings", userop.patch_bonus_notification_settings, methods=["PATCH"], response_class=JSONResponse),
     APIRoute("/member/roles/history/{historyid}", userop.delete_role_history, methods=["DELETE"], response_class=JSONResponse),
     APIRoute("/member/resign", userop.post_resign, methods=["POST"], response_class=JSONResponse),